% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_object.R
\name{agent}
\alias{agent}
\alias{agent,rRofexConnection-method}
\title{See Agent}
\usage{
agent(x)

\S4method{agent}{rRofexConnection}(x)
}
\arguments{
\item{x}{S4 Class. rRofexConnection object}
}
\value{
Scalar with the 'agent'
}
\description{
Shows information about the agent set with \code{\link{trading_login}}
}
