% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_object.R
\name{base_url}
\alias{base_url}
\alias{base_url,rRofexConnection-method}
\title{See Base URL}
\usage{
base_url(x)

\S4method{base_url}{rRofexConnection}(x)
}
\arguments{
\item{x}{S4 Class. rRofexConnection object}
}
\value{
Scalar with the 'base url'
}
\description{
Shows information about the 'base url' where the user has been connected with \code{\link{trading_login}}
}
