% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helpers.R
\name{.validate_fecha}
\alias{.validate_fecha}
\title{Helper: Date validation}
\usage{
.validate_fecha(date)
}
\arguments{
\item{date}{Date}
}
\value{
TRUE if date has a correct format.
}
\description{
Validate date
}
