% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_object.R
\name{login_date_time}
\alias{login_date_time}
\alias{login_date_time,rRofexConnection-method}
\title{See Log-in Timestamp}
\usage{
login_date_time(x)

\S4method{login_date_time}{rRofexConnection}(x)
}
\arguments{
\item{x}{S4 Class. rRofexConnection object}
}
\value{
Scalar with the 'log-in timestamp'
}
\description{
Shows information about the connection timestamp when calling \code{\link{trading_login}}
}
