% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_api.R
\name{primary_login}
\alias{primary_login}
\title{Primary API LogIn}
\usage{
primary_login(username = "ahassel731", password = "hheqvK5<",
  environment = "reMarkets")
}
\arguments{
\item{username}{User Name}

\item{password}{Password}

\item{environment}{String. Wich environment are you going to connect
#'\itemize{
\item reMarkets: Testing environment.For credentials go to \url{https://remarkets.primary.ventures/}
}}
}
\value{
If correct, it will save a token into the current environment
}
\description{
\code{primary_login} log in the user into de Primary API
}
\examples{
\dontrun{primary_login(username="pepe", password="pepino", environment="reMarkets")}
}
