% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRofex.R
\docType{package}
\name{rRofex-package}
\alias{rRofex}
\alias{rRofex-package}
\title{rRofex: Interface to ROFEX APIs}
\description{
Provides a convenient wrapper for consuming data from ROFEX APIs: Trading API, Risk API and BackOffice API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matbarofex}
  \item Report bugs at \url{https://github.com/matbarofex/issues}
}

}
\author{
\strong{Maintainer}: Augusto Hassel \email{ahassel@primary.com.ar}

}
