% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rRofex_connection}
\alias{rRofex_connection}
\title{Create rRofex Connection Object}
\usage{
rRofex_connection(token, base_url)
}
\arguments{
\item{token}{String. **Mandaroty** Obtained with \code{\link{trading_login}}}

\item{base_url}{String. **Mandaroty** URL given by  \code{\link{trading_login}} or known by the client.}
}
\value{
Creates a 'rRofex_connection' S4 Object
}
\description{
\code{rRofex_connection} creates a New Connection Object.
}
\note{
You can use accesors to get information about the Object by using:
\itemize{
\item \code{token(conn)}
\item \code{base_url(conn)}
\item \code{login_date_time(conn)}
}
}
