% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_object.R
\name{rRofexConnection-method}
\alias{rRofexConnection-method}
\alias{token}
\alias{token<-}
\alias{base_url}
\alias{base_url<-}
\alias{login_date_time}
\alias{token,rRofexConnection-method}
\alias{token<-,rRofexConnection-method}
\alias{base_url,rRofexConnection-method}
\alias{base_url<-,rRofexConnection-method}
\alias{login_date_time,rRofexConnection-method}
\title{rRofexConnection-method}
\usage{
token(x)

token(x) <- value

base_url(x)

base_url(x) <- value

login_date_time(x)

\S4method{token}{rRofexConnection}(x)

\S4method{token}{rRofexConnection}(x) <- value

\S4method{base_url}{rRofexConnection}(x)

\S4method{base_url}{rRofexConnection}(x) <- value

\S4method{login_date_time}{rRofexConnection}(x)
}
\arguments{
\item{x}{S4 Class. rRofexConnection object}

\item{value}{New value to be assigned}
}
\description{
rRofexConnection-method
}
