% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_object.R
\name{show,rRofexConnection-method}
\alias{show,rRofexConnection-method}
\title{Show summary of rRofexConnection}
\usage{
\S4method{show}{rRofexConnection}(object)
}
\arguments{
\item{object}{S4 Class. rRofexConnection object}
}
\value{
Summary text with User, Environment and Timestamp
}
\description{
Shows a summary about the rRofexConnection object created with \code{\link{trading_login}}
}
