% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_object.R
\name{token}
\alias{token}
\alias{token,rRofexConnection-method}
\title{See Token}
\usage{
token(x)

\S4method{token}{rRofexConnection}(x)
}
\arguments{
\item{x}{S4 Class. rRofexConnection object}
}
\value{
Scalar with token
}
\description{
Shows information about the token thas has been generated with \code{\link{trading_login}}
}
