% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trading_account}
\alias{trading_account}
\title{Account Information}
\usage{
trading_account(connection, account, detailed = FALSE)
}
\arguments{
\item{connection}{S4. \strong{Mandatory} Formal rRofexConnection class object}

\item{account}{String. \strong{Mandatory} Account Number}

\item{detailed}{Logical. Expanded information.}
}
\value{
If correct, it will load a tibble.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}
Access information about the trading account.
}
\seealso{
Other account functions: 
\code{\link{trading_account_report}()}
}
\concept{account functions}
