% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trading_account_report}
\alias{trading_account_report}
\title{Account Report}
\usage{
trading_account_report(connection, account)
}
\arguments{
\item{connection}{S4. \strong{Mandatory} Formal rRofexConnection class object}

\item{account}{String. \strong{Mandatory} Account Number}
}
\value{
If correct, it will load a tibble.
}
\description{
Access report about your trading account.
}
\note{
To access nested data is strongly recommended the use of `pluck`.
}
\examples{
\dontrun{
data \%>\% pluck("detailedAccountReports", 1, "availableToOperate", 1, "cash")
}
}
\seealso{
Other account functions: 
\code{\link{trading_account}()}
}
\concept{account functions}
