% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trading_currencies}
\alias{trading_currencies}
\title{Currencies}
\usage{
trading_currencies(connection)
}
\arguments{
\item{connection}{S4. \strong{Mandaroty} Formal rRofexConnection class object}
}
\value{
If correct, it will load a data frame.
}
\description{
Access currencies prices.
}
\seealso{
Other market data functions: 
\code{\link{trading_mdh}()},
\code{\link{trading_md}()}
}
\concept{market data functions}
