% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_api.R
\name{trading_instruments}
\alias{trading_instruments}
\title{Primary API Insturments}
\usage{
trading_instruments(request, environment = "reMarkets",
  sec_detailed = FALSE)
}
\arguments{
\item{request}{The type of request that you are making:
\itemize{
\item segments: Available Market Segments
\item securities: Available Instruments listed on Rofex
}}

\item{environment}{String. Wich environment are you going to connect:
\itemize{
\item reMarkets: Testing environment
}}

\item{sec_detailed}{Logical. Optional for environment=securities. Brings aditional information like segment, price, minimal/maximal trading quantity, settlement date, etc.}
}
\value{
If correct, it will load a data frame.
}
\description{
\code{trading_instruments} list segments and instruments currently available in Primary API.
}
\examples{
\dontrun{trading_instruments()}
}
