% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trading_instruments_fronts}
\alias{trading_instruments_fronts}
\title{Front Month of Futures}
\usage{
trading_instruments_fronts(connection)
}
\arguments{
\item{connection}{S4. \strong{Mandatory} Formal rRofexConnection class object}
}
\value{
If correct, it will load a tibble data frame
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("stable")}
List all front month contracts for futures.
}
\seealso{
Other reference data functions: 
\code{\link{trading_instruments}()}
}
\concept{reference data functions}
