% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_api.R
\name{trading_login}
\alias{trading_login}
\title{Primary API LogIn}
\usage{
trading_login(username, password, environment = "reMarkets")
}
\arguments{
\item{username}{User Name}

\item{password}{Password}

\item{environment}{String. Wich environment are you going to connect
\itemize{
\item reMarkets: Testing environment.For credentials go to \url{https://remarkets.primary.ventures/}
}}
}
\value{
If correct, it will save a token into the current environment
}
\description{
\code{trading_login} log in the user into de Primary API
}
\examples{
\dontrun{trading_login(username="pepe", password="pepino", environment="reMarkets")}
}
