% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trading_lookup}
\alias{trading_lookup}
\title{Lookup Order Status}
\usage{
trading_lookup(connection, lookup_type, order_id, proprietary)
}
\arguments{
\item{connection}{S4. **Mandaroty** Formal rRofexConnection class object}

\item{lookup_type}{String. Look-up by:
\itemize{
\item COID. Client Order ID.
\item OID. Order ID. (Not Available)
}}

\item{order_id}{String. ID given by the \code{trading_orders} method.}

\item{proprietary}{String. ID given by the \code{trading_orders} method.}
}
\value{
A data frame.
}
\description{
The method \code{trading_lookup} is used to check the satus of an order.
}
