% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_api.R
\name{trading_md}
\alias{trading_md}
\title{Primary API Market Data Real Time}
\usage{
trading_md(environment = "reMarkets", market_id = "ROFX", symbol,
  entries = c("BI", "OF", "LA", "OP", "CL", "SE", "OI"), depth = 1L)
}
\arguments{
\item{environment}{String. Wich environment are you going to connect:
\itemize{
\item reMarkets: Testing environment
}}

\item{market_id}{String. Market to wich we are going to connect.
\itemize{
\item ROFX. Matba/Rofex
}}

\item{symbol}{String. Use \code{\link{primary_instruments}} to see which symbols are available.}

\item{entries}{Vector of Strings. It contains the information to be required:
\itemize{
\item BI. Bid.
\item OF. Offer.
\item LA. Last Available Price.
\item OP. Open Price.
\item CL. Close Price.
\item SE. Settlement Price.
\item OI. Open Interest.
}}

\item{depth}{Ineteger. Depth of the book to be retrivied.}
}
\value{
If correct, it will load a data frame.
}
\description{
\code{trading_md} retrivies Market Data in Real Time.
}
\examples{
\dontrun{trading_md(symbol='I.RFX20')}
}
