% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_api.R
\name{trading_mdh}
\alias{trading_mdh}
\title{Primary API Historical Market Data}
\usage{
trading_mdh(environment = "reMarkets", market_id = "ROFX", symbol,
  date, date_from, date_to)
}
\arguments{
\item{environment}{String. Wich environment are you going to connect:
\itemize{
\item reMarkets: Testing environment
}}

\item{market_id}{String. Market to wich we are going to connect.
\itemize{
\item ROFX. Rofex: Rosario Futures Exchange.
\item MATBA. Matba: Mercado a Termino de Buenos Aires.
}}

\item{symbol}{String. Use \code{\link{primary_instruments}} to see which symbols are available.}

\item{entries}{Vector of Strings. It contains the information to be required:
\itemize{
\item BI. Bid.
\item OF. Offer.
\item LA. Last Available Price.
\item OP. Open Price.
\item CL. Close Price.
\item SE. Settlement Price.
\item OI. Open Interest.
}}

\item{depth}{Ineteger. Depth of the book to be retrivied.}
}
\value{
If correct, it will load a data frame.
}
\description{
\code{trading_mdh} retrivies Historical Trades for a given instrument.
}
\examples{
\dontrun{trading_mdh(symbol='I.RFX20')}
}
