% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_api.R
\name{trading_new_order}
\alias{trading_new_order}
\title{Send Order to the Market}
\usage{
trading_new_order(symbol, side, quantity, price, order_type = "Limit",
  time_in_force = "Day", account)
}
\arguments{
\item{symbol}{String. Use \code{\link{primary_instruments}} to see which symbols are available.}

\item{side}{String. Either 'Buy' or 'Sell'.}

\item{quantity}{Numeric. Quantity of the order.}

\item{price}{Numeric. Price of the order.}

\item{order_type}{String. Type of order.
\itemize{
\item Limit. Limit order sets the maximum or minimum price at which you are willing to buy or sell.
\item MLL. Market with Leftover as Limit (market order then unexecuted quantity becomes limit order at last price).
}}

\item{time_in_force}{String. Specifies how long the order remains in effect. Absence of this field is interpreted as 'Day':
\itemize{
\item Day. Day or session.
\item IOC. Immediate or Cancel. (Not Available)
\item FOK. Fill or Kill. (Not Available)
\item GTD. Good Till Date. (Not Available)
}}

\item{account}{String. Account Number / Account ID.}
}
\value{
List with outputs like state of the order.
}
\description{
The method \code{trading_new_order} is use to send orders.
}
