% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_api.R
\name{trading_orders}
\alias{trading_orders}
\title{View Orders}
\usage{
trading_orders(account)
}
\arguments{
\item{account}{String. Account Number / Account ID.}
}
\value{
A data frame.
}
\description{
The method \code{trading_orders} is used to see each order sent by Account.
}
