% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_websocket.R
\name{trading_ws_close}
\alias{trading_ws_close}
\title{WS: Close connection}
\usage{
trading_ws_close(websocket_connection)
}
\arguments{
\item{websocket_connections}{web socket that is going to be closed.}
}
\description{
WS: Close connection
}
\examples{

To close many connections at once

\dontrun{
purrr::walk(.x = as.list(tickers), .f = ~ trading_ws_close(websocket_connection = stringr::str_c("ws_", .x)))
}
}
