% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_websocket.R
\name{trading_ws_orders}
\alias{trading_ws_orders}
\title{Web Sockets: Orders Lookup}
\usage{
trading_ws_orders(
  connection,
  destination,
  account = NA,
  only_active = FALSE,
  where_is_env = .GlobalEnv
)
}
\arguments{
\item{connection}{S4. \strong{Mandatory} Formal rRofexConnection class object}

\item{destination}{String. Name of the tibble where the data is going to be stored.}

\item{account}{List. List of accounts to be listeting}

\item{only_active}{Logical. Wheater or not to listen to only active orders}

\item{where_is_env}{Environment. \strong{Only for advance users}.}
}
\value{
If correct, it will load a tibble.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
This method brings orders states in real time using web socket protocol.
}
\seealso{
Other websocket functions: 
\code{\link{trading_ws_close}()},
\code{\link{trading_ws_md}()}
}
\concept{websocket functions}
