% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_object.R
\name{user_name}
\alias{user_name}
\alias{user_name,rRofexConnection-method}
\title{See User Name}
\usage{
user_name(x)

\S4method{user_name}{rRofexConnection}(x)
}
\arguments{
\item{x}{S4 Class. rRofexConnection object}
}
\value{
Scalar with the 'user_name'
}
\description{
Shows information about the user name connected using \code{\link{trading_login}}
}
