% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_api.R
\name{primary_instruments}
\alias{primary_instruments}
\title{Primary API Insturments}
\usage{
primary_instruments()
}
\description{
\code{primary_instruments} list instruments currently available in Primary API.
}
