% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRofex.R
\docType{package}
\name{rRofex-package}
\alias{rRofex}
\alias{rRofex-package}
\title{rRofex: Interface to 'Matba Rofex' Trading API}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Execute API calls to the 'Matba Rofex' <https://apihub.primary.com.ar> trading platform. Functionality includes accessing account data and current holdings, retrieving investment quotes, placing and canceling orders, and getting reference data for instruments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matbarofex.github.io/rRofex}
  \item \url{https://github.com/matbarofex/rRofex}
  \item Report bugs at \url{https://github.com/matbarofex/rRofex/issues}
}

}
\author{
\strong{Maintainer}: Augusto Hassel \email{mpi-augusto@primary.com.ar}

Other contributors:
\itemize{
  \item Juan Francisco Gomez [contributor]
  \item Matba Rofex [copyright holder]
}

}
