% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRofex.R
\docType{package}
\name{rRofex-package}
\alias{rRofex}
\alias{rRofex-package}
\title{rRofex: Interface to ROFEX APIs}
\description{
This package lets you access ROFEX APIs using R. Having this will enable you to integrate ROFEX data easily into R workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com./matbarofex}
}

}
\author{
\strong{Maintainer}: Augusto Hassel \email{ahassel@primary.com.ar}

}
