% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRofex.R
\docType{package}
\name{rRofex-package}
\alias{rRofex}
\alias{rRofex-package}
\title{rRofex: Interface to Matba Rofex Trading API}
\description{
Provides a convenient wrapper for consuming data from Matba Rofex Trading API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://matbarofex.github.io/rRofex}
  \item \url{https://github.com/matbarofex/rRofex}
  \item Report bugs at \url{https://github.com/matbarofex/rRofex/issues}
}

}
\author{
\strong{Maintainer}: Augusto Hassel \email{ahassel@primary.com.ar}

}
