% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rRofex_connection}
\alias{rRofex_connection}
\title{Create rRofex Connection Object}
\usage{
rRofex_connection(token, base_url)
}
\arguments{
\item{token}{String. \strong{Mandatory} Obtained with \code{\link{trading_login}}}

\item{base_url}{String. \strong{Mandatory} URL given by  \code{\link{trading_login}} or known by the client.}
}
\value{
S4 rRofexConnection object.
}
\description{
\code{rRofex_connection} creates a New Connection Object.
}
\note{
You can use accessors to get information about the Object by using:
\itemize{
\item \code{token(conn)}
\item \code{base_url(conn)}
\item \code{login_date_time(conn)}
}
}
\seealso{
Other connection functions: 
\code{\link{trading_login}()}
}
\concept{connection functions}
