% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_object.R
\name{rRofex_connection}
\alias{rRofex_connection}
\title{Create rRofex Connection Object}
\usage{
rRofex_connection(token, base_url, user_name)
}
\arguments{
\item{token}{String. \strong{Mandatory} Obtained with \code{\link{trading_login}}}

\item{base_url}{String. \strong{Mandatory} URL given by  \code{\link{trading_login}} or known by the client.}

\item{user_name}{character. User Name}
}
\value{
S4 rRofexConnection object.

A valid rRofexConecciont object.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("stable")}
\code{rRofex_connection} creates a New Connection Object.
}
\note{
You can use accessors to get information about the Object by using:
\itemize{
\item \code{token(conn)}
\item \code{base_url(conn)}
\item \code{login_date_time(conn)}
\item \code{agent(conn)}
\item \code{user_name(conn)}
}
}
