% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_object.R
\name{rRofexConnection-method}
\alias{rRofexConnection-method}
\alias{token}
\alias{base_url}
\alias{login_date_time}
\alias{agent}
\alias{user_name}
\alias{token,rRofexConnection-method}
\alias{base_url,rRofexConnection-method}
\alias{login_date_time,rRofexConnection-method}
\alias{agent,rRofexConnection-method}
\alias{user_name,rRofexConnection-method}
\alias{show,rRofexConnection-method}
\title{rRofexConnection-method}
\usage{
token(x)

base_url(x)

login_date_time(x)

agent(x)

user_name(x)

\S4method{token}{rRofexConnection}(x)

\S4method{base_url}{rRofexConnection}(x)

\S4method{login_date_time}{rRofexConnection}(x)

\S4method{agent}{rRofexConnection}(x)

\S4method{user_name}{rRofexConnection}(x)

\S4method{show}{rRofexConnection}(object)
}
\arguments{
\item{x}{S4 Class. rRofexConnection object}

\item{object}{S4 Class. rRofexConnection object}
}
\description{
rRofexConnection-method
}
