% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trading_cancel_order}
\alias{trading_cancel_order}
\title{Cancel Order Sent to the Market}
\usage{
trading_cancel_order(connection, order_id, proprietary)
}
\arguments{
\item{connection}{S4. \strong{Mandaroty} Formal rRofexConnection class object}

\item{order_id}{String. clOrdId given by the \code{trading_orders} method.}

\item{proprietary}{String. ID given by the \code{trading_orders} method.}
}
\value{
List with outputs like state of the order.
}
\description{
The method \code{trading_cancel_order} is use to send orders.
}
\seealso{
Other order placements functions: 
\code{\link{trading_new_order}()}
}
\concept{order placements functions}
