% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trading_cancel_order}
\alias{trading_cancel_order}
\title{Cancel Order Sent to the Market}
\usage{
trading_cancel_order(connection, id, proprietary)
}
\arguments{
\item{connection}{S4. \strong{Mandatory} Formal rRofexConnection class object}

\item{id}{String. \strong{Mandatory} clOrdId given by the \code{trading_orders} method.}

\item{proprietary}{String. \strong{Mandatory} ID given by the \code{trading_orders} method.
\itemize{
\item \strong{PBCP}
}}
}
\value{
If correct, it will load a tibble.
}
\description{
The method \code{trading_cancel_order} should be use to cancel orders that are open on the market.
}
\seealso{
Other order placements functions: 
\code{\link{trading_new_order}()}
}
\concept{order placements functions}
