% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trading_login}
\alias{trading_login}
\title{Primary API Log-In}
\usage{
trading_login(username, password, base_url)
}
\arguments{
\item{username}{String. User Name}

\item{password}{String. Password}

\item{base_url}{String. Wich environment are you going to connect:
\itemize{
\item reMarkets: 'http://api.remarkets.primary.com.ar'
\item production: 'https://api.primary.com.ar'
\item xOMS: 'https://api.<BROKER>.xoms.com.ar'
}

Note that reMarkets works with 'http' not 'https'.}
}
\value{
If correct, it will save a hidden token into the current environment
}
\description{
\code{trading_login} log in the user into de Primary API
}
\note{
\itemize{
\item reMarkets: Testing environment. For credentials go to \url{https://remarkets.primary.ventures}
\item production: Production environment. For credentials send an email to \email{mpi@primary.com.ar}
\item xOMS: Ask your broker about it.
}

Accesors:
\itemize{
\item \code{token(conn)}
\item \code{base_url(conn)}
\item \code{login_date_time(conn)}
}
}
\examples{
\dontrun{conn <- trading_login(username = "pepe", password = "pepino", base_url = "http://api.remarkets.primary.com.ar")}
}
