% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_api.R
\name{trading_login}
\alias{trading_login}
\title{Primary API LogIn}
\usage{
trading_login(username, password, env = "reMarkets")
}
\arguments{
\item{username}{String. User Name}

\item{password}{String. Password}

\item{env}{String. Wich environment are you going to connect
\itemize{
\item reMarkets: Testing environment. For credentials go to \url{https://remarkets.primary.ventures}
\item production: Production environment. For credentials send an email to \email{mpi@primary.com.ar}
}}
}
\value{
If correct, it will save a hidden token into the current environment
}
\description{
\code{trading_login} log in the user into de Primary API
}
\examples{
\dontrun{trading_login(username="pepe", password="pepino", env="reMarkets")}
}
