% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trading_lookup}
\alias{trading_lookup}
\title{Lookup Order Status}
\usage{
trading_lookup(connection, lookup_type, id, proprietary)
}
\arguments{
\item{connection}{S4. \strong{Mandatory} Formal rRofexConnection class object}

\item{lookup_type}{String. \strong{Mandatory}. Look-up by:
\itemize{
\item \strong{COID} - Client Order ID.
\item \strong{OID} - Order ID.
}}

\item{id}{String. \strong{Mandatory}. ID given by the \code{trading_orders} method. Depends on `lookup_type`.}

\item{proprietary}{String. ID given by the \code{trading_orders} method. Only for 'lookup_type=COID' In most cases:
\itemize{
\item \strong{PBCP}
}}
}
\value{
If correct, it will load a tibble.
}
\description{
The method \code{trading_lookup} is used to check the status of an order.
}
\seealso{
Other order management functions: 
\code{\link{trading_orders}()}
}
\concept{order management functions}
