% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_api.R
\name{trading_lookup}
\alias{trading_lookup}
\title{Lookup Order Status}
\usage{
trading_lookup(lookup_type, order_id, proprietary)
}
\arguments{
\item{lookup_type}{String. Look-up by:
\itemize{
\item COID. Client Order ID.
\item OID. Order ID. (Not Available)
}}

\item{order_id}{String. ID given by the \code{trading_order} method.}

\item{proprietary}{String. ID given by the \code{trading_order} method.}
}
\value{
A data frame.
}
\description{
The method \code{trading_lookup} is used to check the satus of an order.
}
