% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trading_md}
\alias{trading_md}
\title{Primary API Market Data Real Time}
\usage{
trading_md(
  connection,
  market_id = "ROFX",
  symbol,
  entries = c("BI", "OF", "LA", "OP", "CL", "SE", "OI"),
  depth = 1L
)
}
\arguments{
\item{connection}{S4. \strong{Mandaroty} Formal rRofexConnection class object}

\item{market_id}{String. Market to wich you are going to connect.
\itemize{
\item ROFX. Matba Rofex
}}

\item{symbol}{String. Use \code{\link{trading_instruments}} to see which symbols are available.}

\item{entries}{Vector of Strings. It contains the information to be required:
\itemize{
\item BI. Bid.
\item OF. Offer.
\item LA. Last Available Price.
\item OP. Open Price.
\item CL. Close Price.
\item SE. Settlement Price.
\item OI. Open Interest.
}}

\item{depth}{Integer. Depth of the book to be retrivied.}
}
\value{
If correct, it will load a data frame.
}
\description{
\code{trading_md} retrivies Market Data in Real Time.
}
\seealso{
Other market data functions: 
\code{\link{trading_mdh}()}
}
\concept{market data functions}
