% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_api.R
\name{trading_mdh}
\alias{trading_mdh}
\title{Primary API Historical Market Data}
\usage{
trading_mdh(market_id = "ROFX", symbol, date, date_from, date_to)
}
\arguments{
\item{market_id}{String. Market to wich we are going to connect.
\itemize{
\item ROFX. Rofex: Rosario Futures Exchange.
\item MATBA. Matba: Mercado a Termino de Buenos Aires.
}}

\item{symbol}{String. Use \code{\link{primary_instruments}} to see which symbols are available.}

\item{date}{String. Date to be queried. With format '\%Y-\%m-\%d'.}

\item{date_from}{String. Used together with 'date_to'.}

\item{date_to}{String. Userd together with 'date_from'.}
}
\value{
If correct, it will load a data frame.
}
\description{
\code{trading_mdh} retrivies Historical Trades for a given instrument.
}
\examples{
\dontrun{trading_mdh(symbol='I.RFX20')}
}
