% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trading_orders}
\alias{trading_orders}
\title{View Orders}
\usage{
trading_orders(connection, account)
}
\arguments{
\item{connection}{S4. \strong{Mandatory} Formal rRofexConnection class object}

\item{account}{String. \strong{Mandatory} Account Number}
}
\value{
If correct, it will load a tibble.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("stable")}
The method \code{trading_orders} is used to see each order sent by Account.
}
\seealso{
Other order management functions: 
\code{\link{trading_lookup}()}
}
\concept{order management functions}
