% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trading_orders}
\alias{trading_orders}
\title{View Orders}
\usage{
trading_orders(connection, account)
}
\arguments{
\item{connection}{S4. \strong{Mandaroty} Formal rRofexConnection class object}

\item{account}{String. Account Number / Account ID.}
}
\value{
A data frame.
}
\description{
The method \code{trading_orders} is used to see each order sent by Account.
}
\seealso{
Other order management functions: 
\code{\link{trading_lookup}()}
}
\concept{order management functions}
