% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_websocket.R
\name{trading_ws_close}
\alias{trading_ws_close}
\title{Web Sockets: Close connection}
\usage{
trading_ws_close(close_all = TRUE, selection)
}
\arguments{
\item{close_all}{Logical. Should all connections be closed or only the selected ones.}

\item{selection}{List. Is the same name that you have chosen for destination in \code{\link{trading_ws_md}}}
}
\value{
If correct, it will show a message saying that the connection has been closed.
}
\description{
This method it is use to close open Websocket connections.
}
\examples{

# To close all connections at once

\dontrun{
trading_ws_close(close_all = TRUE)
}
}
\seealso{
Other websocket functions: 
\code{\link{trading_ws_md}()},
\code{\link{trading_ws_orders}()}
}
\concept{websocket functions}
